+++ DarwinStebs/DarwinStebs/Stebs/Opcodes/ASMOperation.cs
@@ -5,7 +5,7 @@ namespace DarwinStebs
 {
 	public class ASMOperation
 	{
-		public int OpCode{ get; set;}
+		public byte OpCode{ get; set;}
 		public string Name { get; set;}
 
 		public List<ASMParameterType> Parameter {get; set;}
@@ -15,19 +15,19 @@ namespace DarwinStebs
 			Parameter = new List<ASMParameterType> ();
 		}
 
-		public ASMOperation(int opCode, string name) : this()
+		public ASMOperation(byte opCode, string name) : this()
 		{
 			OpCode = opCode;
 			Name = name;
 		}
 
-		public ASMOperation(int opCode, string name, ASMParameterType param1)
+		public ASMOperation(byte opCode, string name, ASMParameterType param1)
 			: this(opCode, name)
 		{
 			Parameter.Add (param1);
 		}
 
-		public ASMOperation(int opCode, string name, ASMParameterType param1, ASMParameterType param2)
+		public ASMOperation(byte opCode, string name, ASMParameterType param1, ASMParameterType param2)
 			: this(opCode, name, param1)
 		{
 			Parameter.Add (param2);
